+++ CHANGELOG.md
@@ -19,6 +19,8 @@ interface to `module:oauth2orize.RequestProcessor`.
 interface to `module:oauth2orize.RequestParametersProcessor`.
 - Renamed `http://i.authnomicon.org/oauth2/authorization/http/ResponseMode`
 interface to `module:oauth2orize.Responder`.
+- Renamed `http://i.authnomicon.org/oauth2/authorization/http/ResponseParameters`
+to `module:oauth2orize.responseParametersFn`.
 
 ## [0.0.4] - 2021-12-13
 ### Added